Added ssl heads and deleted linear.py
